requirements/dev.txt
wheel
pylint
pylint-flask
pylint_flask_sqlalchemy
pytest
